fix ethernet handler data structure, and divide thrusster control function into 2 segment.

+++ firmware/sensor_handler.h
@@ -1,8 +1,6 @@
 #ifndef DEPTH_SENSOR_H
 #define DEPTH_SENSOR_H
 
-
-
 #define SENSOR_CONSTANT 234
 
 #include "hardware_config.h"
@@ -10,6 +8,6 @@
 #include <Adafruit_ADS1X15.h>
 
 void init_sensor();
-int16_t read_depth_sensor();
+float read_depth_sensor();
 
 #endif
